+++ lib/src/utils/blend_icons.dart
@@ -1,10 +1,7 @@
 import 'package:flutter/material.dart';
 
-/// Enum for list of icons
-enum BlendIcons { rection, comment, share }
-
 /// Bicons contains a list of [IconData] that the app uses
-class BIcons {
+class BlendIcons {
   /// <i class="material-icons md-36">favorite</i> &#x2014; material icon named "favorite".
   ///
   static const IconData favorite =
